+++ src/Exceptions/RouterException403.php
@@ -3,7 +3,7 @@ namespace GCWorld\Routing\Exceptions;
 
 use Exception;
 use GCWorld\Routing\Hook;
-use GCWorld\Routing\RouterExceptionInterface;
+use GCWorld\Routing\Interfaces\RouterExceptionInterface;
 
 /**
  * Class RouterException403