+++ src/Interfaces/HandlerInterface.php
@@ -1,0 +1,15 @@
+<?php
+namespace GCWorld\Routing\Interfaces;
+
+interface HandlerInterface
+{
+    /**
+     * @return string
+     */
+    public function getTitle();
+
+    /**
+     * @return void
+     */
+    public function setBreadcrumbs();
+}