+++ src/Interfaces/RawRoutesInterface.php
@@ -1,5 +1,5 @@
 <?php
-namespace GCWorld\Routing;
+namespace GCWorld\Routing\Interfaces;
 
 /**
  * Interface RawRoutesInterface